requirements.txt
python>=3.11
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0